+++ packages/blocto-sdk/src/lib/responseSessionGuard.ts
@@ -2,11 +2,11 @@ import { removeItem } from './localStorage';
 
 export default async function responseSessionGuard<T>(
   response: Response,
-  instance: { code: string | null; sessionKey: string }
+  instance: { session: { code: string | null }; sessionKey: string }
 ): Promise<T> {
   if (response.status === 403) {
     // eslint-disable-next-line
-    instance.code = null;
+    instance.session.code = null;
     removeItem(instance.sessionKey);
     throw new Error('[Blocto SDK]: Session expired!');
   }