+++ packages/blocto-sdk/src/main.ts
@@ -3,10 +3,7 @@ import SolanaProvider from './providers/solana';
 import AptosProvider from './providers/aptos';
 import { BloctoSDKConfig } from './index.d';
 
-// eslint-disable-next-line
-import dotenv from 'dotenv';
-
-dotenv.config();
+const sharedSession = { connected: false, code: null, accounts: {} };
 
 export default class BloctoSDK {
   ethereum?: EthereumProvider;
@@ -15,13 +12,25 @@ export default class BloctoSDK {
 
   constructor({ appId, ethereum, solana, aptos }: BloctoSDKConfig) {
     if (ethereum) {
-      this.ethereum = new EthereumProvider({ ...ethereum, appId });
+      this.ethereum = new EthereumProvider({
+        ...ethereum,
+        appId,
+        session: sharedSession,
+      });
     }
     if (solana) {
-      this.solana = new SolanaProvider({ ...solana, appId });
+      this.solana = new SolanaProvider({
+        ...solana,
+        appId,
+        session: sharedSession,
+      });
     }
     if (aptos) {
-      this.aptos = new AptosProvider({ ...aptos, appId });
+      this.aptos = new AptosProvider({
+        ...aptos,
+        appId,
+        session: sharedSession,
+      });
     }
   }
 }